Configure 5 buckets for next Cirrus AB test
This does not start the test, it only sets up the configuration
that will be used by the test. Test will run on 19 wikis
with new MLR models comparing classic (pre-mlr), a model
trained with a minimal hand picked feature set, and a model
trained with a featureset chosen via mRMR feature selection.

Bug: T187148
Change-Id: I745d8c7a1da50fa1a681a07294f49e7f73b29d35