Rollout VirtualPageViews (final stage)
As the 4th and final part of the roll out plan we'll
turn this on for all wikis.

Given English and German Wikipedia do not have page previews
enabled, this change the majority of new traffic to the endpoint
will come from Spanish, Italian, Chinese, Polish and portuguese.

Traffic from enwiki and dewiki will be minor until launch of
the feature

Bug: T189906

commit 948eafb0d50c4cad505701963f8d673800a76ed9
Author: Roan Kattouw <roan.kattouw@gmail.com>
Date:   Tue Mar 27 16:21:53 2018 +0200

Change-Id: Iba670c6cf8ff7df7954bf4d5d8abe23b4e7e7601