Rollout VirtualPageViews (final stage)
As the 4th and final part of the roll out plan we'll
turn this on for all wikis.

Given English and German Wikipedia do not have page previews
enabled, this change the majority of new traffic to the endpoint
will come from Spanish, Italian, Chinese, Polish and portuguese.

Traffic from enwiki and dewiki will be minor until launch of
the feature

Bug: T189906
Change-Id: I380b0c0c77aab8253ba33591de2d4eb93e3f55b8