Enable $wgAbuseFilterProfile on every wiki
Profiling was disabled since it was slowing down edits. However, since
then there have been an improvement in how stats are recorded and the
problem was resolved. This means that we should be able to safely re-enable profiling everywhere. This is for the $wgAbuseFilterProfile part, which should be pretty straightforward.

Bug: T191039
Change-Id: I135d232885a5ff62800d1992a7e4f2a5952cca0a