Update patch set 1
Patch Set 1:

As I understand it, wfLoadExtension queues an extension to be loaded. It does not itself load an extension (this differs from the old days of PHP require_once). This is good and bad.

Good: It means extensions are loaded at a controlled point in time, and in a cacheable and performant way.

Good 2: It means configuration must happen before loading, and configuration cannot depend on ExtensionRegistry.

Bad: It means when you call wfLoadExtension() and ask isLoaded() after, it does not (yet) return true. As such, unless we confirm this with a test, my guess would be that this code does not work for it never returns true, even when it does load before Echo. (ergo: The condition is unreachable).

Patch-set: 1