Update patch set 3
Patch Set 3:

I will create a second backup of the 3rd cluster, with that we will have a recent offline copy of all of es. This could be done regularly ("only" takes 12 hours and no more than 4TB for each the rw ones) until we have a fully automatic process.

Patch-set: 3