Disable writing wb_terms search fields on Wikidata
We disabled *reading* from these fields two weeks ago, in commit
e78328eab0 (change I1fec4c66e2), but kept writing to them for a while in
case we needed to roll back the reading change in case of unexpected
problems. No such problems have been reported, so it should be okay to
stop using the columns altogether, for both reading and writing.

This was already the case for testwikidata, so we can now merge the
configuration sections for testwikidata and wikidata and have the same
behavior on all production systems. (In fact, labs has the same config
too, so we could even merge this into Wikibase.php… perhaps later.)

Bug: T189777
Change-Id: I3938c4cc0729dd55beb2d48efdb2de1b04a85dc2