Update patch set 1
Patch Set 1: Code-Review+1

In hewiki WP:VP there is a consensus for enabling Related Articles, with ability of users to disable it in preferences. Guycn2 suggested to create a CSS gadget to allow users to hide it. 

AFAIK there is no built in configuration in the RelatedArticles extension to allow selection in preferences whether to load it or not. under this assumption this patch is OK and the extra load of hidden parts is not a big deal from performance perspective. 

However, if there is an option in the extension configuration (e.g if I missed it) to allow users to select it in their preferences, please add it too.

Patch-set: 1
Reviewer: Eranroz <228@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1