Lower CirrusSearch delayed job drop to 2 hours
A recent incident of freezing writes for longer than
expected caused jobs to backup into the job queue. We would
have expected the queue size to level off when we hit
the drop timeout, but it seems there were too many jobs
to run.

Help the system out by reducing the timeout from 3 hours
to 2. A separate patch is also doubling the minimum job
delay so there is less work to churn through as well.

Change-Id: I67e395fb592a856d508960662843c4f7a9bb7ce1