multiversion: Remove unused vendor/autoload from getMWVersion.
Follows-up 435506d7 and 8dfd5639c which removed use of CDB
for multiversion, which was the only reason it was loading vendor.

The only reason we still have a vendor is for wmf-config (which loads
after multiversion), where profiler.php uses Mongo/XHGui interfaces.

The require to ensure that is in MWMultiVersion.php, and we can
move that at a later time, but getMWVersion is a standalone script
that definitely doesn't need it.

Change-Id: I5f75c93f8ac545a1dde415911f25b011b361bd2b