Update patch set 1
Patch Set 1:

I've done a plain git-mv in this commit, but that might not actually work. There's a couple of references in vendor/composer that try to find the baseDir. I've re-run composer update in the next commit, but to avoid a broken intermediary commit, might have to squash these two, and sacrifice the clean diff.

Patch-set: 1