Move /multiversion/vendor to /vendor
Follows-up from 7c1245f52c and related commits from 1-2 years ago.
The libraries provided by vendor in wmf-config no longer relate
in any way to multiversion. Rather, they relate to other files
in wmf-config (specifically, onlt profiler.php at this point, but
that may change in the future).

Therefore, remove the special-casing of having vendor within this
subdirectory and place it back to where it would be by default
in a repository using composer, which is in /vendor.

Done as follows to make sure composer is aware of the move:

* Updated composer.json
* Ran `composer install --no-dev` (uses the lock file)
* git rm -r --cached multiversion/vendor/
* git add vendor/

Looks like all changes in vendor are recognised as clean file
moves by Git, and it seems Composer didn't need to change anything
in its autoloader files with .. or dirname. Yay.

Change-Id: I779d2ee01eae5d841579e385f1f9a62b00995352