Update patch set 2
Patch Set 2:

I thought about this, and at first I thought it could work- now I think it will not. Also moving it to m5 was a bad idea.

Note that my request was to either productionize it and make it as good as any other host, including firewall and production configuration, or remove it from production. Now I think a test host, that it has access to production hardware and passwords is a really bad idea. I didn't ask you to move anything to m5, I asked you to either remove it from configuration, OR set it up as well as the others- but it shouldn't live on the same mysql instance as m5, if it is to be hosted on codfw. All m5 services have to be on the same database service.

Note I am not agains the idea of it living on m5, doing that solves the scope of T192339, but as you may have seen, you will not be able to write locally to codfw. The solution you chose has issues. If it were to live on codfw and m5, it must use TLS and connection pooling. I would recommend a separate instance local to codfw, but well configured. If it is part of mediawiki, which means it can be served from eqiad, it cannot be exposed also without tls and connection pooling. This is a WMF policy -and literally a Q4 goal related to multi-dc- for all intra-dc communication (it has to be encrypted, and encrypted connections are normally too slow to be done without connection pooling).

Patch-set: 2