Update patch set 2
Patch Set 2:

thanks for catching this. PHP will cast that value to string anyway so it shouldn't have any effect. The config change is a string as previously Popups was available as a beta feature (BetaFeatures extension) and we tried to keep same ON|OFF values for both enabled as a beta feature and enabled as a regular feature. 
I changed it to be a string so we have same types everywhere.

Patch-set: 2
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0