Enable testing LanguageConverter in sandboxes on deploymentwiki
We enable the Translate extension on deploymentwiki, which sets
$wgPageLanguageUseDB as a side-effect, thus giving `admin`s and
`translationadmin`s on deploymentwiki the ability to manually
set a page's language (using [[Special:PageLanguage]]) in order
to test LanguageConverter (and its Parsoid port).

For good measure, we also set $wgUsePigLatinVariant, which ensures
that the default language on deploymentwiki (en) also has a variant
defined (en-x-piglatin), so LanguageConverter can be tested on
English-language articles on deploymentwiki.

Bug: T143628
Change-Id: Ia830ca1ad71a73cd77ff132476580e771268c805