Stop loading the MwEmbedSupport extension, part I
This code has moved into the TimedMediaHandler extension directly
from version 1.32.0-wmf.10 onwards. Once we're sure we won't ever
regress from wmf.10, this can be safely deployed as effectively a
no-op. (Not deploying it shouldn't have any impact on production,
as the code was written to be backwards-compatible).

This is part 1 of 4, to keep in line with the "only patches which
touch one file at a time" deployment policy.

Change-Id: Ic5b8648ab93e40f26bca74045c164da953b4fec5