Scap clean: remove remote cache directory
Currently "scap clean" removes the local cache directory and then
attempts to sync with all remotes; however, we exclude a subdirectory of
cache (l10n) from rsync since it is generated on each host.  Since we
exclude that directory rsync --delete ignores that directory and leaves
it behind after the sync. The result is that deployers see this problem
on debug hosts when they run "scap pull":

    cannot delete non-empty directory: php-1.XX.0-wmf.X

And old versions leave their l10n directories on hosts until manual
cleanup. This removes the directory directly on hosts as part of: "scap
clean".

Bug: T157030
Change-Id: Iaa91830491839245fcc594f0f1b5b39813ace95a