Beta Cluster: Increase wgNavigationTimingSamplingFactor from 0.1% to 10%
The Beta Cluster has significantly less traffic, so for the metrics
to be at least somewhat serve as test data for deployment-webperf11,
increase the sampling a bit.

With prod's sample rate, we typically collect about 250 data
points per minute in production. In Beta Cluster, in the 27 days
that the service has been enabled, it has collecte all of 19 data
points. With entire days going by without a single data point.

Bumping this up 100x should be fine, and is quite conservative still.

Bug: T195314
Change-Id: I573489660d0773161d696528a5f4ceff942591df