Update patch set 5
Patch Set 5: Code-Review-1

In a later patch is fine, and per Aaron, might even be obsolete if it ends up removed indeed. If the key name is the only difference then I don't mind. Just want to make sure the rest is as similar as possible between prod and beta so that we actually catch issues there. 

Speaking of which, blocking this until T198280 is resolved, which we found after a9124c2f5867, which applied to beta the same as proposed here for mw.org.

Patch-set: 5
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1