Update patch set 2
Patch Set 2:

Okay, I think I still want to separate s/Tidy/RemexHtml/g with cleanup of the tidy config ... so that if there is something we overlook in the tidy config cleanup, we can revert that to fix it and not rollback RemexHtml.

Anomie: To answer your specific question, I didn't think ParserMigration would be broken -- I just wanted us to make sure we don't break it in any way since that also uses tidy config.

We want to support it for a few months after Tidy is removed so that editors can continue using it (if it helps them) to fix pages that have changed rendering after the RemexSwitch. A few months down the line, we can disable Tidy completely and repurpose ParserMigration for other changes down the line.

Patch-set: 2
Reviewer: Subramanya Sastry <257@e9e9afe9-4712-486d-8885-f54b72dd1951>