Update patch set 1
Patch Set 1:

wgRightsUrl is actually used as a default for the "dataRightsUrl" setting WikibaseRepo uses for the little JS popup messages that appear when you start editing anonymously. I'm not sure if *this* particular value is used there. Probably not, because of the missing protocol. Otherwise a broken link would appear in these popups.

As far as I can see "dataRightsUrl" is properly set and not derived from wgRightsUrl on the cluster. Please make sure this is the case.

Patch-set: 1
CC: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>