services: Convert LabsServices.php to static array file
Makes it more clear that it must not depend on any global variables
and must not have any conditionals or other complex statements.

This is also in preparation of making it safe to include outside
MediaWiki initialisation. For example, in profiler.php from PhpAutoPrepend
which runs before any of multiversion, MediaWiki or wmf-config.

Change-Id: I2bff4eff4eb33b176454e20ce1f4409c9f5a1391