services: Convert ProductionServices.php to static array file
Previously done with LabsServices.php.

This makes it more clear that the file must not depend on any global
variables. This is also in preparation of making it safe to include outside
MediaWiki initialisation. For example, in profiler.php from PhpAutoPrepend
which runs before any of multiversion, MediaWiki or wmf-config.

Change-Id: I4db5d03f8af7fd22210eeb55feea3a2f13fc211b