Don't use include_once for assigned values ($wgInterwikiCache)
include_once and require_once return bool(true) when called
a second time for the same file, which would be a disaster
in this case.

Instead, use `include`. This matches what we do for other
static files that return values in wmf-config and MediaWiki core
(e.g. $db in missing.php, or Resources.php in core).

Change-Id: I9bd529ce8a2bd5078e965f426542450c651d1cc6