Update patch set 1
Patch Set 1:

Jdlrobson - that's correct, but as Olga said, we want that as a default behavior no matter what. Popups are visible to anons by default, if you create an account we also want to make it visible by default. If in the future we decide to enable PagePreviews on different projects we will most probably want to keep the same behavior, because of that I decided to set it to ON for all possible projects.

Patch-set: 1
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>