Update patch set 1
Patch Set 1:

I read Olga's comment to mean all wikis that have the feature enabled already. Note we haven't spoken to any of the projects that don't have this feature. 

I'm guessing/hoping that when we deploy to other wikis we will remember this is the default and factor that as part of our roll out plan. Definitely worth making this clear to Olga so that when we roll out to wikidata we are thinking about this.

Anyway I'm fine with this in case that wasn't clear but wanted to point out the default entry is redundant on a bunch of wikis that this feature is not enabled on.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>