search.wikimedia.org should properly handle multivalue separation char (0x1F)
If we send a param value starting with 0x1F on a non multi value param, the
MW-API will immediately emit an error with badvalue_notmultivalue.
Rather than trying to map all MW-API errors to 400 vs 500 we detect this bad
search request early and save a call to the backend.

Change-Id: I59350b8e5f262a83f54e1347d3918da53440573b