Delete multiversion/submodules.json, putatively unused
It's my contention that this file, however much dutiful people like
myself add and remove entries to it, is not actually used by anything.
I've come to this conclusion based on the lack of edits to it by
others as they add and remove files. Consequently, I propose deleting
it and seeing what happens. Let's hope I'm right, and that this file
is left behind from refactorings of the scap and deploy tools. Ish.

Change-Id: I109b8bfc2bb3b748f704a311fd7efc571d50d88e