Switch in WebM VP9/Opus video transcodes to replace WebM VP8/Vorbis
Changes TimedMediaHandler's ffmpeg threading and resource limits
suitably for the newer VP9 configuration, which packs higher quality
into less bits for most files, and much higher quality in more bits
for files with high frame rates or lots of detail/motion.

* enable VP9 transcodes, disable VP8 ones
* change from 2 to 8 ffmpeg threads per job
* enable row-based multithreading (requires new ffmpeg)
* increase resource limits for higher compute time

Once this is deployed, new uploads of videos will produce transcodes
in WebM VP9/Opus instead of VP8/Vorbis. Existing transcodes will
continue to play back and be used transparently, until such time
as they get cleaned up by a batch process or update to the file.

If it's necessary to undo the VP9 switch, comment out the portion
that modifies $wgEnabledTranscodeSet and let it use the defaults.

Bug: T63805
Change-Id: Iae6d2ed9fd0736d6bfad09509d685acf9d57d696
Depends-On: I550749e8f6897a789773c6924d39886010d5f625