Update patch set 3
Patch Set 3:

AIUI this would make the request go through Varnish, and the custom domain was used to avoid that. Not sure how important that is. To disable the request entirely, wgUploadThumbnailRenderMap should be commented out instead. (It just fetches thumbnails of freshly uploaded files in standard sizes, to avoid long loading time for the first user who would look at them; it's no big deal to disable. Especially since apparently it did not work in production for the last two years.)

Patch-set: 3