Re-enable VP8 video transcodes to fix playback regression
With the VP8 transcodes off we not only weren't generating new ones,
but the old ones weren't getting used for playback on new page
renderings, even if no VP9 replacement was yet available.

Re-enabling them will cause double-encoding of new uploads, but
this is a relatively small cost compared to the batch re-encode
to new VP9 versions and is fairly harmless.

Change-Id: Icc36f09bf237510ed5be1c9fbb8eeddd2781c641