Don't use hex escapes for non-ASCII characters
Outside the vendor directory, there are only two instances in PHP files
where hexadecimal escape sequences are used, both for the UTF-8 bytes of
the letter "è" (U+00E8 LATIN SMALL LETTER E WITH GRAVE), and both within
wmf-config/InitialiseSettings.php. For consistency with the rest of the
file, unescape this character.

Change-Id: I3776676b162292a115bccd17c8be93642c4ab5d4