Update patch set 1
Patch Set 1: Code-Review-1

This patch can not be deployed as-is. There is a (recent) rule (will find it) that a patch must be deployed with one command, and your patch requires two commands. It should be easy to split it in two patches, one changing the logos, one using the new logos. Let me know if you need help with that.

Patch-set: 1
Reviewer: Zfilipin <487@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1