Update patch set 1
Patch Set 1: Code-Review-1

Sorry, but it needs to be the other way around. Currently this patch removes the reference, and is set as child of the parent commit that removes the creation of the variable. That would mean CommonSettings.php starts emitting errors after the first patch lands.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1