Make current default of MCR migration stage explicit.
We need the current default to be explicit in the live config, so we don't inadvertently
change it for the live systems when we change the default for master and CI with T198561.

This config change effectively does nothing, it just insulates the live system against 
changes to DefaultSettings.php.

Bug: T197816
Change-Id: Ib718868d2c768b6ea851355eef047bc0e6593495