Add editsitejson to everyone who has editinterface
The permission for editing JSON pages in the MediaWiki namespace
was separated from editinterface, so non-admin groups have
lost that permission. This patch fixes that. It's a low-risk
permission so it's not worth making communities go through some kind
of request process, easier to just give it to everyone who had it
in the past. (Except for some groups that obviously do not need it like
centralnotice-admin and votewiki interface translators.)

Bug: T190015
Change-Id: I7900beb86d2543d1dfb6fed987899943b3fa70b8