Test that all wikis are in one of the section dblists
We already have a test that all.dblist contains every wiki in any
dblist, so we only need to test that everything in all.dblist is also in
one of the section dblists (s1 through s8 or wikitech).

Do the same for the size dblists, since we can easily enough.

Bug: T202904
Change-Id: I43c79297499ccb1787626df25fe8e8fe64848bc3