Update patch set 1
Patch Set 1:

The commit can not be deployed as-is during SWAT because it changes files in two root folders (static and wmf-config). Please split the commit in two, the first one adding files to static folder, the other one changing the file in wmf-config to use the new logos.

Patch-set: 1
CC: Zfilipin <487@e9e9afe9-4712-486d-8885-f54b72dd1951>