Add high-density logos for Commons
This commit adds two high-density logos for Commons
based on the logo currently in use on the project.

It also modifies the current low-density logo used
by the wiki to avoid mismatches between the standard
version and the new logos.

There is a tiny difference between the current logo
and the new one as the SVG version isn't exactly
the same as the original logo created in PNG in 2004,
but the change is almost imperceptible, so let's give
it a go.

Bug: T203342
Change-Id: I3b6afb14233e785c3b9960354ea794df61d87195