Update patch set 2
Patch Set 2:

The task mentioned 'retry-after'. I think using sleep would be fine, if we're worried about adding complexity to CP.

However, if the added feature in CP isn't a problem, then I think it would be better to use 'retry-after' so that we're not occupying web request resources on the jobrunner servers.

CP on the other hand, being asynchronous, could use a timeout or some other part of its loop, without needing to actually occupy any significant resources.

Patch-set: 2
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>