Update patch set 2
Patch Set 2:

Afaik 'retry-after' doesn't, or shouldn't, relate to job loss (retry exhaustion during read-only).  Because as you say, "retrying after X time" can still be within the read only period.

We prevent job loss by making the change prop handler not decrement the allowed  number of retries if the response is 'readonly': true, which also should apply to jobs that allow 0 retries, it wasn't a real try basically.

As for database contention, I would expect the job runner to be able to respond very quickly to read-only without even hitting the DB. MediaWiki does the same. Firstly, in longer cases of read-only, it'd be statically configured, so no db hit. And in the case of natural read-only (due to high lag), the information is cached and shared between requests afaik.

Anyhow, I'm not sure I understand why a 45s sleep in the web request is different from a 45s retry-after response, except that the former occupies a significant resources where the latter doesn't. I'm probably just missing something, just don't know what it is.

Patch-set: 2