Update patch set 2
Patch Set 2:

The difference stems from the what change-prop is implemented. Originally, we thought to effectively pause the CP workers during the read-only period (cf. T204154#4583876), but that is not easily implementable. So, the alternative is to re-enqueue these jobs. Because of that, we need a way to control the effective request rate produced by CP, and the quickest thing to do currently is to sleep here. While this does hold up jobrunner resources, it does not greatly impact the outcome during the read-only period since 99% of jobs need the DB anyway.

Patch-set: 2
Reviewer: Mobrovac <2177@e9e9afe9-4712-486d-8885-f54b72dd1951>