logging: Disable 'Wikibase.NewItemIdFormatter' channel
Being spammed for over a week with "CRITICAL" alerts that are
clearly not critical.

Normally when a channel is broken/spammy, we lower the
entry-level threshold, but this is not possible because the
abused level is 'CRITICAL', which is the highest severity.

Feel free to revert once the error has lowered its severity to
no higher than WARNING, given it is a recoverable and caught
exception. In particular, do not revert if the current instance
of the issue has been fixed, because it must not cause more
false alarms again in the future.

Bug: T204791
Change-Id: Icab62550c8728acb232bf4e0656aec5289787693