Update patch set 1
Patch Set 1:

The plan is to move them for now on eqiad only, and reuse the read only time for the datacenter switch back to move the configuration so no extra read only time is needed.

This will make eqiad and codfw out of sync, but we will make sure to do the appropiate changes before and after to make things not break (with the experience of the wikidata/s8 split gave us).

Patch-set: 1
Reviewer: Jcrespo <2556@e9e9afe9-4712-486d-8885-f54b72dd1951>