profiler: Prevent Xenon flushes from being able to fatal a request
Since the switch from Eqiad to Codfw, latency to mwlog1001 has
naturally increased. Where normally we lose under 10 samples
per month, we now lose about 10 samples per week.

That's still a very acceptable loss. The timeout for Redis was
intentionally set low so that it's a best-effort approach whilst
being very low in terms of potential risk for congestion or load.
If for whatever reason it can't flush it quickly, we bail.

The good thing is while connect() always degrades by gracefully
by returning false and writing a non-fatal warning to the logs.

The ba news is that, publish() isn't as graceful, it throws if
the socket read after any publish command reaches the time out.

Bug: T206092
Change-Id: If607ad111a23197a98045225316806452fb04f1e