profiler: Move flush to a function and prep beta/prod consolidate (beta-only)
This will share the code between beta and prod instead of duplicating
in two places. It is also a step toward having different flush in
PHP7 vs HHVM.

For now, this uses the new file (which is otherwise logically identical)
only Beta Cluster. Once it works fine there for a couple hours, the next
commit applies it to production.

Bug: T176916
Change-Id: Id79baae9071ccef64af9ce362ef14d327de9e05b