import: Fix broken labs-only config in wmfImportSources()
This never worked.

This code was introduced with Ib325df7c76837e1, which merged Dec 2015.
Another change, I62b46314, merged before that in Oct 2015, which moved
the dblists files to a subdirectory. Hence this logic was always
selecting the production files, regardless of realm.

This is fine, however, because there are no DB or realm-specific
versions of these files. Hence, the call to getRealmSpecificFilename()
is redundant.

Change-Id: I5f7ff4234bcdee6202292d43b60b68c9b71278ae