multiversion: Remove reference to deprecated getRealmSpecificFilename()
This is not used anywhere and new usage has been discouraged
since 2015. I'm keeping the function as a one-liner for back-compat
so that the change can be safely deployed.

Once rolled out and confirmed no usage in any currently deployed
code, including private files, the next commit will remove it.

Bug: T189966
Change-Id: Ice9a74e734816d5c1505293fac5dc6d8803f0582