Refactor wmf-config/env.php out of multiversion/MWRealm.php
Unlike MWRealm, this file doesn't set global variables and isn't
meant to be included via 'require_once'. Instead, it exports a
static array that is meant to be included via 'require'.

This makes it usable in standalone entry points that don't (or
can't) make guarantees about whether MediaWiki is or will be
initialised, such as fatar-error.php and profiler.php.

We should also remove $wmfCluster, which isn't used anywhere in
Wikimedia Git, nor anywhere under deploy1001:/srv/mediawiki-staging
or mwdebug1001:/srv/mediawiki. But I'll leave that for a later
commit.

Change-Id: I79fb3d194a58556ee7b7b383d908ac43297a43b6