Update patch set 2
Patch Set 2:

Please leave a note to deployers if merge something to mediawiki-config or make sure it gets deployed (to mwdeploy1001 at least, the rest probably doesn't matter for a beta-only patch). Otherwise it will show up in the diff when the next time someone does a deploy and cause confusion.

Patch-set: 2
CC: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>