Add item & property to wmgWikibaseClientRepoNamespaces for wiktionaries
This issue was introduced in I8589b90cef748

Previously whenever $wmgUseWikibaseClient was set then
$wgWBClientSettings['repoNamespaces'] would be set to include the item
and property namesapces.

The patch switched this to only set repoNamesapces for things in
the wikidataclient db list.
The wikidataclient db list for some reason only includes
enwiktionary...

T208317 has been filed as a followup to clean this up...

Bug: T208293
Change-Id: I312d8642cc2df992d28b920913b14910da4cd72d