Update patch set 2
Patch Set 2:

This will just give a cryptic error when someone tries to use it (there are links to the page from request templates); that's OK if we are in a hurry but not great. Maybe we could have a generic DisabledSpecialPage which just displays a message, stash the message in WikimediaMessages, and override the normal special page class here?

Patch-set: 2
CC: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>