BC: Enable Schema.org page split test
Add target namespace(s) and buckets in Wikibase.php as they won't differ
between the Beta Cluster and production. For both completeness and
safety, set a default sampling rate to 0 which will disable the test
unless the pageSchemaSplitTestSamplingRate Wikibase client setting is
set explicitly.

Finally, set the sampling rate on the Beta Cluster to 1, meaning that
50% of pages will receive the treatment.

Bug: T208763
Change-Id: I9c5694e8effecce92a839a3d4640d683229eee22