[WIP][DNM][BROKEN] Not finished tests for wg config of WMF wikis
This is the initial step of the experiment to have
some kind of the way of testing that wg global config
variables per wiki are set accordingly.
Reasoning is to have some more certainty that when
one makes change to a line on wmf-config/* files,
it only affects the global(s) one intended to,
and only for the expected wiki(s).

The patch is completely unfinished, and not even close
to work for a single case. I am just dumping the code
I have in case my HDD decided to crash.

Would submit as a non-public draft but it is no longer
possible on Gerrit.

Bug: T208493
Change-Id: If3d66b0017e810deff3534776af15fc506570f11