Set AdvancedSearch to default on group0 wikis
The wmgUseAdvancedSearch flag turns the extension on/off. Currently it's
always on, which means we could even remove this code. But this is not
what the current ticket is about.

wgAdvancedSearchBetaFeature is a config flag the extension understands.
If enabled (which currently is the default), the extension puts itself in
Beta mode, allowing users to opt-in. When disabling this Beta flag, the
extension will be active for all users, and disappear from the Beta
features page.

The relevant line of code for this behavior is:
https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/AdvancedSearch/+/refs/heads/master/includes/Hooks.php#33

Note this patch here will also allow us to change the default in the
codebase.

Planned for the mid-day SWAT on 2018-11-08.

Bug: T207641
Change-Id: I55fa90aed7fbb0faf41e9b5fd3ac5b1e300143d8