Update patch set 2
Patch Set 2: Code-Review-2

This has been extensively discussed in T153563, and ultimately declined. Entity URIs are supposed to be stable, and thanks to HSTS and HSTS preloading, they will effectively be loaded over HTTPS anyways on all modern browsers.

A mismatch between a client repo URL and the repo concept URI should not be a problem – ultimately, the URI is just an identifier, and as far as I’m aware, nothing in Wikibase should ever try to resolve it. (For example, on non-Wikidata Wikibase installations, the entity URI will often still be http://DOMAIN/entity, even if no redirect for /entity is set up in the web server.) If this mismatch is causing problems, then there’s a bug somewhere, but breaking our stable identifiers is not the correct solution, sorry.

Patch-set: 2
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2