Update patch set 1
Patch Set 1: Code-Review+1

In this case I actually think we could go for HTTPS :) stability is not a concern here (because we’re changing the base URI anyways, and also because it’s only the beta cluster), and I’m personally not convinced by Stas’ tradition argument in T153563#2884409 (for me, stability is the only reason to keep HTTP for Wikidata URIs). But that’s not a strong opinion from my side, just an observation.

Patch-set: 1
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1