Update AbuseFilter config to keep the status quo
This is so that in I58f215b8aa3444602532098a3c9dfd25409f29eb (which
depends-on this patch) we'll be able to change some defaults within
AbuseFilter's extension.json without changing the resulting config
for WMF wikis.

NOTE: This patch should be deployed together with the patch in AF to
avoid discontinuities.

Change-Id: I09cb462312c5bbc71ee3a8f0fbd91357908130d4