Update patch set 2
Patch Set 2:

Beta Cluster picks up the prod configuration without the override, doesn't it? So removing it now would drop back to WRITE_NEW.

I'm planning on pushing NEW out to production over the course of the week. This was the first step, then group0, then group1, then group2 will put it everywhere and we can probably remove it in both places to rely on the new default in DefaultSettings.php.

Patch-set: 2