Don’t send SPARQL prefixes in WikibaseQualityConstraints
This configuration setting informs the WikibaseQualityConstraints
extension that the SPARQL endpoint we use, the Wikidata Query Service
(see the WBQualityConstraintsSparqlEndpoint setting, a few lines above),
has special support for Wikibase, which means that we don’t need to send
Wikibase-specific SPARQL prefixes (they’re defined by default).

This should result in a significant reduction in network traffic from
the application servers to the WDQS servers (since the queries WBQC
currently sends are short except for the prefixes, and not many other
things running on the application servers use WDQS, see [1]), but as
that network traffic is not expected to make up a large part of the
traffic of either the application servers or the WDQS servers as a
whole, the effect may not be very noticeable.

[1]: https://wikitech.wikimedia.org/wiki/Wikidata_query_service/Usage

Bug: T204317
Change-Id: I18d8025cbbf5a7f91356772dfc7d6dc99c9782cc