errorpages: Remove unused php-fatal-error.html file
The php5 Puppet module that used to refer to this file no longer
exists as of 491729da8c9e4b1.

The php7 Puppet module does not currently install the php-wmerrors
extension nor does it specify a fatal error handler by other
means.

If and when such a file is needed for PHP 7, it should be created
by Puppet similar to If534171fd59cf70e3 for HHVM.

Bug: T113114
Bug: T201140
BUg: T187147
Change-Id: Ic093c3122f5f048e685134e0539c4781633be122