Allow MediaInfo on commons to reference Wikidata items.
This enables the "federation" feature of WikibaseRepo on Commons,
by declaring "wikidata" as a foreign repository.

Note that this does not change the fact that WikibaseCLient, which is
also active on Commons, uses no prefix (or rather, the empty prefix)
when accessing wikidata. This is to be resolved later, see T211800.

Note that this deferation is one-way: wikidata is not set up to
reference MediaInfo entities on commons.

Bug: T204748
Change-Id: I3c7ca0b3ceba3f68ab0b8e520affd8681cb53654