Configure WikibaseQualityConstraints on Beta
The WikibaseQualityConstraints extension reads constraint definitions
from statements on properties, using a set of entity IDs. The default
IDs refer to production Wikidata, both so that no configuration is
required there and to allow others to import those entities from there,
but to usefully run the extension on Beta Wikidata, we need to configure
the different entity IDs used there. The entities were imported from
production Wikidata into Beta Wikidata using a maintenance script (see
T209957#4823447) and this configuration is the result.

Bug: T209957
Change-Id: If89f20ca1ed501880655b62f48777c9bbdb58bf4