Update patch set 1
Patch Set 1:

No harm in this, but it's a no-op (the list only has 10K passwords, and it's unlikely that will ever change as we have switched to Bloom filters for larger lists and there is a separate check for that). Most of these groups should probably use the large list, and doing both checks is a pointless performance hit. Icc70122fab1b5 tries to clean that up, among other things.

Patch-set: 1
CC: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>