Add new WMCS IP range to $wgRateLimitsExcludedIps
Browser tests are being run on WMCS instances. Wikibase tests targetting
testwikidatawiki would trigger the ratelimit, possibly due to the
instance migrated to a new OpenStack region which comes with a different
IP range.

Add 172.16.0.0/21 IP range to $wgRateLimitsExcludedIps for testwikidatawiki.
Clarify the comment, SauceLabs is a third party and is obviously not on
WMCS, it is thus not being applied the rate limit exception.

Bug: T167432
Change-Id: Id929e8ba9144266942aa735ea7df5e69a9e61eb0