Increase MW -> EventBus service HTTP request timeout.
Currently the default timeout is used, which is 5 seconds.
Given that we have a substantial number of timeouts and they are
quite critical, since the event is completely lost in this case,
it seem reasonable to increase the timeout substantially.

The value of 60 is chosen arbitrary as a reasonable in my opinion
tradeoff between 12x increase in timeout vs 300 seconds default
MultiHTTPClient timeout.

Bug: T204183
Change-Id: I1e49271bb992dc394c7c3a7cd2e31522bef3943a