Update patch set 3
Patch Set 3:

OK. That is useful I think :) I would like to know how it helps the newcomer? As I understand it, untracked files (files not part of git) are ignored by default for most purposes.

I believe, it is often not the newcomer, but the experienced user that is using advanced commands or is otherwise bothered by text output in git-status, and want to hide it.

Patch-set: 3
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>