Update patch set 4
Patch Set 4:

> getSafeReadDB() throws when CA is read only. that might be a problem?

As far as I can see it only throws when you try to get the master DB, which only happens when the central user was constructed with the master flag (not the case here) or $wgCentralAuthUseSlaves is false (not the case for the WMF cluster).

IIRC the catch block was meant for the exception that's today handled in CentralAuthUser::queryAttached().

Patch-set: 4
Reviewer: Gergő Tisza <389@e9e9afe9-4712-486d-8885-f54b72dd1951>