InitialiseSettings.php: Increase parsercache TTL
Initially the keys had a TTL of 30 days, but
due to disk space issues (T167784) it had to
be reduced to 22 days.

We have re-freshed the parsercache hosts (T208383),
so we are ready to go back to 30 days TTL.
However, we want to do a small increase first, to make
sure everything works as expected.

Bug: T210992
Change-Id: I146bb638824e13fa86e5e30aaffc5327b43c055a