Update patch set 1
Patch Set 1:

In particular, zero.wikimedia.org (the internal Zero wiki, as opposed to Wikipedia Zero traffic itself) should probably be made inaccessible first (e.g. from dns or VCL) so that we aren't changing the way the wiki behaves and potentially causing leakage of information or otherwise changing its behaviour/config. I'm sure we could do it, but it seems rather risky and expensive to review; much easier if it's unlinked first and then we can go to down on rm-ing it all :)

Patch-set: 1
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>